graduate/requirements.txt
construct==2.5.3
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-filter==2.1.0
future==0.17.1

pefile==2018.8.8
python-decouple==3.1
python-ptrace==0.9.3
pytz==2018.9
six==1.12.0
